refine import_articles.py. we don't want usmf articles, so let's specify that we don't want usmf articles at the start, when we request data from the hydra api (in this case, instead of doing a blacklist of usmf, we'll do a whitelist of the premium services).
also, rename variable 'jr' to 'article_json', so that it's a little more descriptive.

also, to better approximate 15 articles per ticker, run a counter in the for loop, and increment the counter only after we've called save() on the article. why this instead of something like json_response[:15] in the "for" loop (~line 37)? consider the case where we're in the for loop with ticker FB, and json_response[:15] contains 4 duplicates; in that case, we'd add only 11 articles for FB.