articles view band aid - given that we've increased the number of Articles known to the live environment (satellite.fool.com)... the count of Articles is >20k. !!
cool, but... the current implementation for finding an author's services involves finding all of the Articles linked to that author, and then going through all of those matches and pulling out the service name.  to recap, for each author, we're going to analyze 20k Articles. that's not trivial, and we should refine how&when we generate that set of associated services. the efficient way would follow the Ticker design of a dedicated field for 'services_for_ticker', which is assigned to when new data comes in, and otherwise not re-crunched every time we want to know its value.  we should do something similar to hold the services associated with an author.

until then, a cheap way of avoiding crunching 20k Articles every time we want to see the services of an author is to... limit the Articles we consider to those from within the last 2 years. while we're at it, let's increase the number of articles shown per page, from 50 to 100, to enhance the benefit of the dictionary lookup - we expect many of the authors in the additional 50 to have already been in the first 50, so we'll already have crunched (and have handy) the associated services for that author.