Revert "slim down the articles index so that it uses the datatables.js sort/search"
Why: with the previous (ie non-datatables) implementation, we limited how many articles were returned by the server to 100. with the new implementation, the server returns *all* articles and tries to render *all* articles, as it's the client side where the pagination occurs. at the moment, the new implementation takes way too long to load. we can revisit, consider limiting the number of articles that the server tries to render.

This reverts commit 2fb5d1eba8cb8d1b7986fcfd1767fa4f91467e7d.